moved post related components into a folder
